chore: extract custom transitions in separate util
